Split test_piecewise_cuda_graph.py to optimize CI resource usage
- Split 1 GPU tests into test_piecewise_cuda_graph_1_gpu_a.py (6 tests, ~500s)
  and test_piecewise_cuda_graph_1_gpu_b.py (5 tests, ~500s)
- Move 2 GPU test (Qwen3OmniMOE) to test_piecewise_cuda_graph_2_gpu.py (~200s)
  and reduce tp from 4 to 2
- Remove original test_piecewise_cuda_graph.py from per-commit-4-gpu suite
- Update run_suite.py to reference new test files in appropriate suites

This reduces CI time for 4 GPU tests and better distributes the workload.
