+++ test/srt/run_suite.py
@@ -87,6 +87,8 @@ suites = {
         TestFile("test_original_logprobs.py", 41),
         TestFile("test_page_size.py", 60),
         TestFile("test_penalty.py", 82),
+        TestFile("test_piecewise_cuda_graph_1_gpu_a.py", 500),
+        TestFile("test_piecewise_cuda_graph_1_gpu_b.py", 500),
         TestFile("test_priority_scheduling.py", 130),
         TestFile("test_pytorch_sampling_backend.py", 66),
         TestFile("test_radix_attention.py", 105),
@@ -137,6 +139,7 @@ suites = {
         TestFile("test_dp_attention.py", 350),
         TestFile("test_load_weights_from_remote_instance.py", 72),
         TestFile("test_patch_torch.py", 19),
+        TestFile("test_piecewise_cuda_graph_2_gpu.py", 200),
         TestFile("test_eagle_dp_attention.py", 200),
     ],
     "per-commit-4-gpu": [
@@ -145,7 +148,6 @@ suites = {
         TestFile("test_local_attn.py", 411),
         TestFile("test_multi_instance_release_memory_occupation.py", 64),
         TestFile("test_pp_single_node.py", 500),
-        TestFile("test_piecewise_cuda_graph.py", 1260),
         TestFile("test_epd_disaggregation.py", 400),
     ],
     "per-commit-8-gpu-h200": [
